refactor: consolidate duplicate OuraClient implementations
- scripts/alerts.py: Import OuraClient from oura_api.py
- scripts/weekly_report.py: Import OuraClient from oura_api.py
- scripts/telegram_bot.py: Import OuraClient from oura_api.py
- scripts/oura_api.py: Added get_recent_sleep() and get_weekly_summary() methods
  to support telegram_bot.py functionality

Fixes:
- get_recent_sleep() now uses 'score' key (not 'readiness') to match original
- get_recent_sleep() returns last N entries (matching original behavior)
- Fixed sys.path to use resolve().parent for correct import path

Net reduction: ~37 lines (94 removed, 57 added)
